# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAS Sneakers Geo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  description: string;
  imageUrl: string;
  category: string;
  inStock: boolean;
  featured: boolean;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  brand: string;
  price: number;
  description: string;
  imageUrl: string;
  category: string;
  inStock: boolean;
  featured: boolean;
}

export interface AdminVerifyInput {
  password: string;
}

export interface AdminVerifyResponse {
  token: string;
  success: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}