# lib/api-zod/src/generated/types/adminVerifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAS Sneakers Geo API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVerifyResponse {
  token: string;
  success: boolean;
}