# lib/api-zod/src/generated/types/createProductInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAS Sneakers Geo API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductInput {
  name: string;
  brand: string;
  price: number;
  description: string;
  imageUrl: string;
  category: string;
  inStock: boolean;
  featured: boolean;
}